+++ CHANGELOG.md
@@ -33,6 +33,24 @@ Classify the change according to the following categories:
 - Added `summary_by_chunk` endpoint and `job\views.summary_by_chunk`
 - Added `unlink` endpoint and `job\views.unlink` along with **UserUnlinkedRuns**
 
+## Develop - 2023-03-06
+### Minor Updates
+### Added
+- In **REoptjlMessageOutputs** added a `has_stacktrace` field to denote if response has a stacktrace error or not. Default is False.
+
+## Develop 03/08/2023
+### Minor Updates
+##### Changed
+- Updated the following default values to job/ app (v3):
+   - PV, Wind, Storage, CHP, Hot Water Storage, Cold Water Storage, Electric Storage: **federal_itc_fraction(PV,Wind,CHP)** and **total_itc_fraction(Hot Water Storage, Cold Water Storage, Electric Storage)** to 0.3 (30%)
+   - PV, Wind, Storage, CHP, Hot Water Storage, Cold Water Storage, Electric Storage: ***macrs_bonus_fraction** to 0.8 (80%)
+   - Hot Water Storage and Cold Water Storage: **macrs_option_years** to 7 years
+   
+## Develop - 2023-03-02
+ ##### Changed
+ - In `reo/nested_inputs.py` v2 inputs (`defaults_dict[2]`), updated the following default values:
+   - ColdTES, HotTES: **macrs_option_years** to 7 (years)
+   - ColdTES, HotTES: ***macrs_bonus_pct** to 0.8 (80%)
 ## Develop 11/21/2022
 ### Minor Updates
 ##### Added